refactor: route Prod-side toRepresentation through ProdDispatch
Apply the same rename-and-throw pattern as the Sum side. Each Prod typegen's
`toRepresentation` override calls `ProdDispatch.dispatcherBypass` to surface
any caller that bypassed the dispatcher; the body lifts into ProdDispatch.scala
as a private helper (`prodCaseImpl`, `prodCaseUplcConstrImpl`,
`prodCaseOneElementImpl`). ProductCaseUplcOnly's one-line forwarder folds
inline into the outer dispatcher case.

`SumDispatch.toRepresentation`'s typegen match adds a Prod-source case that
delegates to ProdDispatch.toRepresentation — Sum/Prod routing always goes
through SumDispatch.toRepresentation as the entry, then crosses to
ProdDispatch when the source typegen is product-shaped (works around the
SIRType.isSum/isProd trap where sum-shaped constructors like List.Cons carry
CaseClass SIR types).

`ProductCaseOneElementSirTypeGenerator.argLoweredValue` is promoted from
private to package-private so the dispatcher's `prodCaseOneElementImpl`
helper can access it given the gen instance.
